requirements.txt
streamlit==1.35.0
torch==2.3.0
numpy==1.24.4
faiss-cpu==1.7.4
Pillow==10.3.0
ftfy==6.2.0
regex==2023.12.25
tqdm==4.66.4
clip-anytorch==2.5.2